# specs/2-ai-book-creation/contracts/book-api-contract.yaml
openapi: 3.0.0
info:
  title: AI/Spec-Driven Book Creation API
  description: API for managing AI-assisted book creation workflow
  version: 1.0.0
servers:
  - url: https://api.book-creation.example.com/v1
    description: Production server
  - url: https://staging-api.book-creation.example.com/v1
    description: Staging server

paths:
  /books:
    get:
      summary: List all books
      description: Retrieve a list of all books with optional filtering
      parameters:
        - name: status
          in: query
          description: Filter books by status (draft, review, published, archived)
          schema:
            type: string
            enum: [draft, review, published, archived]
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  total:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      summary: Create a new book
      description: Create a new book based on a feature specification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'

  /books/{bookId}:
    get:
      summary: Get book details
      description: Retrieve detailed information about a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update book details
      description: Update information about a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a book
      description: Permanently delete a book and all its content
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
      responses:
        '204':
          description: Book deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /books/{bookId}/chapters:
    get:
      summary: List chapters in a book
      description: Retrieve all chapters for a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
        - name: status
          in: query
          description: Filter chapters by status
          schema:
            type: string
            enum: [draft, review, published]
      responses:
        '200':
          description: List of chapters
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chapter'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Create a new chapter
      description: Create a new chapter in a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterCreateRequest'
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /books/{bookId}/chapters/{chapterId}:
    get:
      summary: Get chapter details
      description: Retrieve detailed information about a specific chapter
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
        - name: chapterId
          in: path
          required: true
          description: Unique identifier of the chapter
          schema:
            type: string
      responses:
        '200':
          description: Chapter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update chapter details
      description: Update information about a specific chapter
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
        - name: chapterId
          in: path
          required: true
          description: Unique identifier of the chapter
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterUpdateRequest'
      responses:
        '200':
          description: Chapter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /books/{bookId}/sections:
    post:
      summary: Create a new section
      description: Create a new section in a specific chapter
      parameters:
        - name: bookId
          in: path
          required: true
          description: Unique identifier of the book
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectionCreateRequest'
      responses:
        '201':
          description: Section created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Section'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /content-blocks/{contentBlockId}/validate:
    post:
      summary: Validate content block
      description: Run validation checks on a content block
      parameters:
        - name: contentBlockId
          in: path
          required: true
          description: Unique identifier of the content block
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                validation_types:
                  type: array
                  items:
                    type: string
                    enum: [content_accuracy, apa_citation, technical_fact, build_integrity]
                  description: Types of validation to perform
      responses:
        '200':
          description: Validation results
          content:
            application/json:
              schema:
                type: object
                properties:
                  validation_results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /research-items:
    get:
      summary: List research items
      description: Retrieve a list of research items with optional filtering
      parameters:
        - name: bookId
          in: query
          description: Filter by associated book
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, in_progress, completed, discarded]
      responses:
        '200':
          description: List of research items
          content:
            application/json:
              schema:
                type: object
                properties:
                  research_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResearchItem'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      summary: Create a research item
      description: Create a new research item for concurrent research
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchItemCreateRequest'
      responses:
        '201':
          description: Research item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchItem'
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the book
        title:
          type: string
          description: Title of the book
        description:
          type: string
          description: Brief description of the book content
        version:
          type: string
          description: Current version of the book
        status:
          type: string
          enum: [draft, review, published, archived]
          description: Current status of the book
        created_at:
          type: string
          format: date-time
          description: Timestamp when book was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update
        authors:
          type: array
          items:
            type: string
          description: List of authors contributing to the book
        target_audience:
          type: string
          description: Intended audience for the book

    BookCreateRequest:
      type: object
      required:
        - title
        - target_audience
      properties:
        title:
          type: string
          description: Title of the book
        description:
          type: string
          description: Brief description of the book content
        target_audience:
          type: string
          description: Intended audience for the book
        feature_spec:
          type: string
          description: Feature specification to base the book on

    BookUpdateRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the book
        description:
          type: string
          description: Updated description of the book content
        status:
          type: string
          enum: [draft, review, published, archived]
          description: Updated status of the book
        target_audience:
          type: string
          description: Updated target audience for the book

    Chapter:
      type: object
      required:
        - id
        - book_id
        - title
        - order
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the chapter
        book_id:
          type: string
          description: Reference to the parent book
        title:
          type: string
          description: Title of the chapter
        order:
          type: integer
          description: Sequential order within the book
        status:
          type: string
          enum: [draft, review, published]
          description: Current status of the chapter
        word_count:
          type: integer
          description: Number of words in the chapter
        module_type:
          type: string
          enum: [ros2, gazebo_unity, isaac_sim, vla, capstone]
          description: Type of module this chapter covers
        learning_objectives:
          type: array
          items:
            type: string
          description: List of objectives for the chapter
        created_at:
          type: string
          format: date-time
          description: Timestamp when chapter was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update

    ChapterCreateRequest:
      type: object
      required:
        - title
        - order
      properties:
        title:
          type: string
          description: Title of the chapter
        order:
          type: integer
          description: Sequential order within the book
        module_type:
          type: string
          enum: [ros2, gazebo_unity, isaac_sim, vla, capstone]
          description: Type of module this chapter covers
        learning_objectives:
          type: array
          items:
            type: string
          description: List of objectives for the chapter

    ChapterUpdateRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the chapter
        order:
          type: integer
          description: Updated sequential order within the book
        status:
          type: string
          enum: [draft, review, published]
          description: Updated status of the chapter
        module_type:
          type: string
          enum: [ros2, gazebo_unity, isaac_sim, vla, capstone]
          description: Updated type of module this chapter covers

    Section:
      type: object
      required:
        - id
        - chapter_id
        - title
        - order
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the section
        chapter_id:
          type: string
          description: Reference to the parent chapter
        title:
          type: string
          description: Title of the section
        order:
          type: integer
          description: Sequential order within the chapter
        content_type:
          type: string
          enum: [text, code, diagram, quiz, lab]
          description: Type of content in the section
        content:
          type: string
          description: Markdown content of the section
        concepts_covered:
          type: array
          items:
            type: string
          description: List of concepts covered in the section
        created_at:
          type: string
          format: date-time
          description: Timestamp when section was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update

    SectionCreateRequest:
      type: object
      required:
        - title
        - order
        - content_type
      properties:
        title:
          type: string
          description: Title of the section
        order:
          type: integer
          description: Sequential order within the chapter
        content_type:
          type: string
          enum: [text, code, diagram, quiz, lab]
          description: Type of content in the section
        content:
          type: string
          description: Markdown content of the section
        concepts_covered:
          type: array
          items:
            type: string
          description: List of concepts covered in the section

    ContentBlock:
      type: object
      required:
        - id
        - section_id
        - type
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the content block
        section_id:
          type: string
          description: Reference to the parent section
        type:
          type: string
          enum: [text, code, image, table, video, quote]
          description: Type of content
        content:
          type: string
          description: The actual content
        metadata:
          type: object
          description: Additional information (alt text, caption, etc.)
        validation_status:
          type: string
          enum: [verified, needs_review, invalid]
          description: Status of validation
        created_at:
          type: string
          format: date-time
          description: Timestamp when content block was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update

    Citation:
      type: object
      required:
        - id
        - content_block_id
        - type
        - apa_format
        - created_at
      properties:
        id:
          type: string
          description: Unique identifier for the citation
        content_block_id:
          type: string
          description: Reference to the content block using the citation
        type:
          type: string
          enum: [book, journal, website, documentation, research_paper]
          description: Type of source
        apa_format:
          type: string
          description: Full APA-formatted citation
        url:
          type: string
          format: uri
          description: URL of the source if applicable
        accessed_date:
          type: string
          format: date
          description: Date the source was accessed
        verified:
          type: boolean
          description: Whether the source was verified
        created_at:
          type: string
          format: date-time
          description: Timestamp when citation was created

    ResearchItem:
      type: object
      required:
        - id
        - book_id
        - topic
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the research item
        book_id:
          type: string
          description: Reference to the associated book
        status:
          type: string
          enum: [pending, in_progress, completed, discarded]
          description: Current status of the research
        topic:
          type: string
          description: Topic or concept to research
        source_url:
          type: string
          format: uri
          description: URL of the research source
        summary:
          type: string
          description: Summary of the research findings
        relevance_score:
          type: number
          minimum: 0
          maximum: 1
          description: Score indicating how relevant the research is
        verified:
          type: boolean
          description: Whether the information was verified
        created_at:
          type: string
          format: date-time
          description: Timestamp when research item was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update

    ResearchItemCreateRequest:
      type: object
      required:
        - book_id
        - topic
      properties:
        book_id:
          type: string
          description: Reference to the associated book
        topic:
          type: string
          description: Topic or concept to research
        source_url:
          type: string
          format: uri
          description: URL of the research source if known

    ValidationResult:
      type: object
      required:
        - rule_id
        - result
        - severity
      properties:
        rule_id:
          type: string
          description: ID of the validation rule
        result:
          type: boolean
          description: Whether validation passed
        severity:
          type: string
          enum: [error, warning, info]
          description: Severity of the validation result
        message:
          type: string
          description: Description of the validation result
        details:
          type: object
          description: Additional details about the validation

    ValidationError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details

  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'